move user list to user controller
