+++ rundeckapp/src/test/groovy/rundeck/controllers/MenuControllerSpec.groovy
@@ -58,7 +58,6 @@ import spock.lang.Specification
 import spock.lang.Unroll
 
 import javax.servlet.http.HttpServletResponse
-import java.awt.Frame
 
 /**
  * Created by greg on 3/15/16.
@@ -1262,120 +1261,6 @@ class MenuControllerSpec extends Specification {
 
     }
 
-    def "loadUsersList summary with last exec"() {
-        given:
-        UserAndRolesAuthContext auth = Mock(UserAndRolesAuthContext)
-        controller.frameworkService=Mock(FrameworkService){
-            1 * getAuthContextForSubject(_)>>auth
-            1 * authorizeApplicationResourceType(_,_,_) >> true
-        }
-        controller.userService.frameworkService = Mock(FrameworkService){
-            getRundeckFramework() >> Mock(IFramework){
-                getPropertyLookup() >> Mock(IPropertyLookup){
-                    hasProperty(UserService.SESSION_ID_ENABLED) >> true
-                    getProperty(UserService.SESSION_ID_ENABLED) >> true
-                }
-            }
-        }
-        def userToSearch = 'admin'
-        def email = 'test@test.com'
-        def text = '{email:\''+email+'\',firstName:\'The\', lastName:\'Admin\'}'
-        def lastSessionId = 'exampleSessionId01'
-        User u = new User(login: userToSearch, lastSessionId: lastSessionId)
-        u.save()
-
-        ScheduledExecution job = new ScheduledExecution(
-                jobName: 'blue',
-                project: 'AProject',
-                groupPath: 'some/where',
-                description: 'a job',
-                argString: '-a b -c d',
-                workflow: new Workflow(
-                        keepgoing: true,
-                        commands: [new CommandExec(
-                                [adhocRemoteString: 'test buddy', argString: '-delay 12 -monkey cheese -particle']
-                        )]
-                ),
-                retry: '1'
-        )
-        job.save()
-        def exec = new Execution(
-                scheduledExecution: job,
-                dateStarted: new Date(),
-                dateCompleted: null,
-                user: userToSearch,
-                project: 'AProject'
-        ).save()
-
-        when:
-        def users = controller.loadUsersList()
-
-        then:
-        response.json.users
-        response.json.users.find{it.login == userToSearch}
-        response.json.users.find{it.login == userToSearch}.lastJob
-        response.json.users.find{it.login == userToSearch}.lastSessionId == lastSessionId
-
-    }
-
-    def "loadUsersList summary with logged in status"() {
-        given:
-        UserAndRolesAuthContext auth = Mock(UserAndRolesAuthContext)
-        controller.frameworkService=Mock(FrameworkService){
-            1 * getAuthContextForSubject(_)>>auth
-            1 * authorizeApplicationResourceType(_,_,_) >> true
-        }
-        controller.userService.frameworkService = Mock(FrameworkService){
-            getRundeckFramework() >> Mock(IFramework){
-                getPropertyLookup() >> Mock(IPropertyLookup){
-                    hasProperty(UserService.SESSION_ID_ENABLED) >> true
-                    getProperty(UserService.SESSION_ID_ENABLED) >> true
-                }
-            }
-        }
-
-        def userToSearch = 'admin'
-        def email = 'test@test.com'
-        def text = '{email:\''+email+'\',firstName:\'The\', lastName:\'Admin\'}'
-        def lastSessionId = 'exampleSessionId02'
-        User u = new User(login: userToSearch, lastSessionId: lastSessionId)
-        u.save()
-
-        ScheduledExecution job = new ScheduledExecution(
-                jobName: 'blue',
-                project: 'AProject',
-                groupPath: 'some/where',
-                description: 'a job',
-                argString: '-a b -c d',
-                workflow: new Workflow(
-                        keepgoing: true,
-                        commands: [new CommandExec(
-                                [adhocRemoteString: 'test buddy', argString: '-delay 12 -monkey cheese -particle']
-                        )]
-                ),
-                retry: '1'
-        )
-        job.save()
-        def exec = new Execution(
-                scheduledExecution: job,
-                dateStarted: new Date(),
-                dateCompleted: null,
-                user: userToSearch,
-                project: 'AProject'
-        ).save()
-
-        when:
-        def users = controller.loadUsersList()
-
-
-        then:
-        response.json.users
-        response.json.users.find{it.login == userToSearch}
-        response.json.users.find{it.login == userToSearch}.lastJob
-        response.json.users.find{it.login == userToSearch}.loggedStatus == UserService.LogginStatus.LOGGEDIN.value
-        response.json.users.find{it.login == userToSearch}.lastSessionId == lastSessionId
-
-    }
 
     def "api job forecast xml"() {
         given:
@@ -1510,61 +1395,4 @@ class MenuControllerSpec extends Specification {
         response.json.futureScheduledExecutions.size() == 1
     }
 
-    def "loadUsersList summary with no session id"() {
-        given:
-        UserAndRolesAuthContext auth = Mock(UserAndRolesAuthContext)
-        controller.frameworkService=Mock(FrameworkService){
-            1 * getAuthContextForSubject(_)>>auth
-            1 * authorizeApplicationResourceType(_,_,_) >> true
-        }
-        controller.userService.frameworkService = Mock(FrameworkService){
-            getRundeckFramework() >> Mock(IFramework){
-                getPropertyLookup() >> Mock(IPropertyLookup){
-                    hasProperty(UserService.SESSION_ID_ENABLED) >> true
-                    getProperty(UserService.SESSION_ID_ENABLED) >> false
-                }
-            }
-        }
-
-        def userToSearch = 'admin'
-        def email = 'test@test.com'
-        def text = '{email:\''+email+'\',firstName:\'The\', lastName:\'Admin\'}'
-        def lastSessionId = 'exampleSessionId02'
-        User u = new User(login: userToSearch, lastSessionId: lastSessionId)
-        u.save()
-
-        ScheduledExecution job = new ScheduledExecution(
-                jobName: 'blue',
-                project: 'AProject',
-                groupPath: 'some/where',
-                description: 'a job',
-                argString: '-a b -c d',
-                workflow: new Workflow(
-                        keepgoing: true,
-                        commands: [new CommandExec(
-                                [adhocRemoteString: 'test buddy', argString: '-delay 12 -monkey cheese -particle']
-                        )]
-                ),
-                retry: '1'
-        )
-        job.save()
-        def exec = new Execution(
-                scheduledExecution: job,
-                dateStarted: new Date(),
-                dateCompleted: null,
-                user: userToSearch,
-                project: 'AProject'
-        ).save()
-
-        when:
-        def users = controller.loadUsersList()
-
-        then:
-        response.json.users
-        response.json.users.find{it.login == userToSearch}
-        response.json.users.find{it.login == userToSearch}.lastJob
-        response.json.users.find{it.login == userToSearch}.loggedStatus == UserService.LogginStatus.LOGGEDIN.value
-        response.json.users.find{it.login == userToSearch}.lastSessionId == null
-
-    }
 }
